refactor: single authority for trip transit allocation
